requirements.txt
easydict
matplotlib
numpy
nuscenes_devkit
open3d
pandas
Pillow
pykitti
pyquaternion
PyYAML
scipy
setuptools
termcolor
timm
tqdm
prodigyopt
xformers # optional if failed, select a version from https://download.pytorch.org/whl/xformers/
einops
# pointnet2_ops: pip install "git+https://github.com/erikwijmans/Pointnet2_PyTorch.git#egg=pointnet2_ops&subdirectory=pointnet2_ops_lib"
# knn_cuda 
# xformers == 0.0.29.post3
# torch == 2.6.0+cu118
# torchvision == 0.21.0+cu118
# cuda: 11.8